#15064. Use directory entry info instead of fopen() when possible, to save CPU

+++ include/mega/logging.h
@@ -407,7 +407,7 @@ public:
                     i++;
                 }
 
-                logger->log(nullptr, level, nullptr, "", dm.get(), dms.get(), i);
+                logger->log(nullptr, level, nullptr, "", dm.get(), dms.get(), static_cast<int>(i));
             }
         }
         for (auto &s: mCopiedParts)
